Improving RandomAlgorithm
Movement comparison of the random method did nothing as there will always be the same amount left in the walletList.

Movement comparison added to the tip cycle instead to account for making change and generating suboptimal changes.

Moved some aspects of the greedy algorithmn into its own function to call.