Fix for SumWalletDebt and DenomCompare
Fix a bug regarding the sumwalletDebt. Now it uses Abs value to avoid being equivalent to the amount paid out to the bank. It will rise when there are large differences between Wallet amounts on the table side.

Also fix bug with comparing denoms. No need to compare anything with that function as movementComparison already stores that every loop.