Updated Jquery
Upgraded to Jquery 2. Jquery 3 breaks Jquery Mobile library so it cannot be used unfortanately.

Changed debtreporting to number instead of text for better gui popups on mobile.

Apparently delegate is depreciated for a LONG time, so I changed it to .on.

Final screen buttons need to be fixed since cannot clikc the tip gui with mouse for some reason.